Changing gatehouse to use containerIp:privatePort
Inter-container communication is a security problem on compute nodes.
While `--icc=true` can prevent that, a hole then still exists because by
default dockerIp:publicPort is accessible by all containers.

The solution is to allow machineshop to set up a container link to the
compute node proxy that does the routing, and then have the proxy
communicate directly to the container.

Changes have been made to dit4c/dit4c-platform-cnproxy to accommodate
this new behaviour.