Improving JS caching and enabling h2 server push
The Application controller observes requests for JavaScript assets that
are not versioned, and sets up RequireJS config statements to make sure
they're used. This works around the problem of not being able to easily
enumerate JS assets at compile time.

This cache of versioned JavaScript is also used to add a "Link" tag
marking them as assets for preload. When used in conjunction with
nghttpx this is used to perform server push of the assets to improve
page load time.